refactor(settings): extract LlamaServerConsole from AISettings
AISettings.tsx was 743 lines, ~590 of them the provider==="localllm" branch:
the on-device Llama server console (model manager, server start/stop, engine
installer/updater, download + binary progress, update check) plus all its
runtime state, lifecycle effect, and ~10 action handlers.

Move that whole self-contained surface into LlamaServerConsole.tsx. It owns its
own llama state; the parent passes contextLimit/onContextLimitChange (persisted
in shared aiConfig) and the shared Max-steps row as a prop so the rendered order
is identical. The local-provider connection-status line moves with the console
(its serverStatus/llamaModels live there). No behaviour or styling change —
markup + className strings copied verbatim.

Host AISettings.tsx drops 743 -> 159 (a clean provider switch). Only AISettings
is imported externally (settings-view), so call-site-safe.

Verified: type-check clean; eslint --max-warnings 0 clean (whole repo); full
suite (1128) green.
